requirements.txt
pyTelegramBotAPI==4.1.1
mercadopago==2.0.8
psycopg2-binary==2.9.2
pytz==2021.3